requirements.txt
python-dotenv
pandas
gradio
langchain_nvidia_ai_endpoints
langchain_community
langchain_core
